@@ -1,9 +1,21 @@
 {
   "name": "chrt-object",
   "version": "0.0.9",
-  "description": "Chart Object components of Chrt",
-  "main": "dist/chrt-object.js",
+  "description": "Universal Object components of Chrt",
+  "type": "module",
+  "main": "src/index.js",
   "module": "dist/chrt-object.esm.js",
+  "jsdelivr": "dist/chrt-object.min.js",
+  "unpkg": "dist/chrt-object.min.js",
+  "exports": {
+    "umd": "./dist/chrt-object.min.js",
+    "default": "./dist/chrt-object.esm.js"
+  },
+  "keywords": [
+    "chrt",
+    "charts",
+    "visualization"
+  ],
   "watch": {
     "build": {
       "patterns": [
@@ -14,10 +26,11 @@
   },
   "scripts": {
     "test": "jest",
-    "postversion": "rollup -c",
     "build": "rollup -c",
+    "postversion": "rollup -c",
     "lint": "eslint src --fix",
-    "watch": "npm-watch"
+    "watch": "npm-watch",
+    "prepare": "husky install"
   },
   "husky": {
     "hooks": {
@@ -36,28 +49,26 @@
   },
   "homepage": "chrt.io",
   "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
+    "@babel/core": "^7.15.8",
+    "@babel/plugin-transform-modules-commonjs": "^7.15.4",
+    "@babel/plugin-transform-runtime": "^7.15.8",
+    "@babel/preset-env": "^7.15.8",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "babel-jest": "^27.0.2",
+    "@rollup/plugin-node-resolve": "^13.0.5",
+    "babel-jest": "^27.2.5",
     "babel-plugin-root-import": "^6.6.0",
     "babelify": "^10.0.0",
     "budo": "^11.6.4",
-    "chrt": "0.0.70",
-    "eslint": "^7.29.0",
-    "husky": "^4.3.0",
-    "jest": "^27.0.4",
-    "npm-watch": "^0.10.0",
+    "chrt": "0.0.105",
+    "eslint": "^7.32.0",
+    "husky": "^7.0.2",
+    "jest": "^27.2.5",
+    "npm-watch": "^0.11.0",
     "rollup": "2",
-    "rollup-plugin-bundle-size": "^1.0.3",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-uglify": "^6.0.4",
-    "watch": "^1.0.2"
+    "watch": "^0.13.0"
   },
   "files": [
     "dist/**/*.js",
